expr/agg: limit intermediate type creation in Schema (#2899)
Each call to Schema.Mixin creates a new record type, but for current use
cases, all but the last are intermediate types that remain in the type
context but never appear in the input.  Creating these intermediate
types can consume substantial CPU and memory.  Rework Schema so it does
not create them.