store config state in a journal and improve index object model (#2931)
This commit changes the storage model for pool/lake configuration
from a json file to a transaction journal.  It also improves the
index abstraction into an index-rules configuration stored as a
journal.

We also updated the lake design docs to reflect this new model
and change reference to data "segment" to "data object".

The pool rename command has been added.