Optional Put
Allow expressions like `x := 1` to get interpreted as a put proc.

The big change here is a change to how the parser handles a list of expressions
without proc (e.g. `to_lower(x), to_upper(y)`). They now get parsed as
an OpExprs proc. It is then on the semantic compiler to look
at the list of expressions and determine if the list of expressions is a
Filter, Summarize or Put proc.

Closes #2850