add branches to lake model (#2976)
This commit adds a branch concept to the zed lake.  This required
some rewiring of the snapshot logic to implement merge and a new
approach to locking across two branches for a merge (though the
locking details are not yet implemented).  The TAIL pointer in a
journal now tracks the base pointer in a parent branch.  TAIL is
transactionally guaranteed through locking (TBD).

For now there is only one root branch (main), and branches must
all be descended from this.

"zed lake query" no longer reports "empty log" as it simply
produces no output.

"zed lake ls" no longer lists objects as meta-queries can be used
for this.  Instead, ls lists pools and branches.

The lake commands "add", "commit", and "status" are deprecated.
The lake command "load" does an automatic commit in the backend.

Closes #2626
Closes #2537
Closes #2975