Limit scope of reserved words (#2985)
Allow IDs to have the name of reserved words (e.g. by, from, where, etc).

In order to make this possible change Where clauses to accept a
LogicalOrExpr instead of BooleanSearch as it was previously doing. The
BooleanSearch made it impossible to distinguish between a keyword such
as by and limit and an ID search primitive.

Closes #2398