simple zson format (#3017)
This commit eliminates the somewhat cryptic integer type IDs from
the ZSON specification and changes the ZSON formatter to output
a record stream where each record is decodable by itself.  This is
done by encoding all needed type definitions in every record.
Originally, we thought this would be too verbose (and sometime it
is --- see below), but the opposite is true... you often would like
to see the types of the fields without having to refer back to the
integer typedef.  Also, ZSON is not meant to be performant so the
extra overhead of encoding all the types is not relevant.

While this is now the default output format, the spec still
allows type definitions to persist across the stream as in ZNG.

Also, an option has been added to the ZSON writer allowing you to
specifying a regular expression where matched type names will be
persisted across the stream.  This will be useful when typedefs
are large and gnarly as happens when fusing records that create
record-subfields that have union type (i.e., a union of different
record types).  In a future PR, we will add an option to fuse
to create typedefs for the record unions with an indicate prefix,
so for example, you could say:

  zq -persist 'shape.*' 'fuse -prefix shape' -