change behavior of lake checkout state (#3024)
This commit changes the default behavior of the checkout command
to store the HEAD pointer in $HOME.  The environment variable
ZED_HEAD_DIR can be defined to override this behavior.  In particular,
setting ZED_HEAD_DIR to ".", will revert to the previous behavior
where the HEAD pointer is stored in the current working directory.