fix sort on fuctions returning internal buffer
To avoid allocations, some Zed function implementations return a
zng.Value whose Bytes field references an internal buffer.  When the
result from such function is used as a sort key, the corresponding
expr.Evaluator is called twice for each call to the expr.CompareFn
returned by expr.NewCompareFn, and because of the internal buffer, the
second call will overwrite the result from the first, ensuring the two
results are always equal.  Fix this by copying the first result before
making the second call.