change "zqd" to "zed" in python client (#3095)
This commit gets rid of the "zqd" reference in the client.
This will break some old examples in online content but we
will have many forthcoming python examples in the near
future.  We also clarified the example in the python
README so newcomers who know python and just want to
quickly try out Zed should have an easier time.