make fuse() output deterministic (#3190)
Output from fuse() is nondeterministic because the order in which
expr/agg.fuse.Result mixes fuse.shapes into its schema depends on map
iteration order, which is unspecified.  Make fuse() output determinstic
by recording the order in which types are added to fuse.shapes and
mixing them into the schema in that order.