enable ZNG compression in lake/api.RemoteSession.Load
"zapi load" is slower than expected because the ZNG writer in
lake/api.RemoteSession.Load does lots of small writes to its
io.PipeWriter.  Speed it up by enabling compression for the ZNG writer.

(Wrapping the io.PipeWriter with a bufio.Writer is would also address
the problem, but we generally don't expect high-entropy data, so
compression should be beneficial.)