lake index.Filter: Use expression tree for lookups (#3197)
First this pr modifies the contents of the index Filter pushdown so
rather than just storing predicates it creates a simplified verision
of binary expression that the index can use.

Secondly the evaluation part of index.Filter is modified so it make use
the expression tree. For instance if the left hand side of an or
operation returns a hit, the expression will return true without having
to wait for the right hand expression to complete. The index.Filter also
uses the index.Finder.Nearest method so compare expressions with the
operators >, >=, <, <= can also be used.

Closes 3183