fix lake/ztests/delete.yaml (#3236)
This applies the change to service/ztests/delete.yaml in #3227 to the
similar test in lake/ztests/delete.yaml.

The test in lake/ztests/delete.yaml is flaky because it expects
"from pool@branch:objects" to emit objects in creation order, but the
actual order is undefined.  Fix it by grabbing the first object ID
immediately after creation.