load -seekStride: Don't set global var
In the zed load command setting the global SeekIndexStride variable to the
value of the -seekStride flag made it so this value was ignored for
service load requests. Instead add seekStride option to the lake Load
command and propagate this value to the object writers. Add seekStride
query parameter for load requests.