Dec 2021 updates to Zed specs (#3319)
This commit updates the Zed specs with lessons learned from 
our past two years developing Zed.

This is a SPEC ONLY update so the code is now out of date with respect
to the specification changes.  This will be remedied in an upcoming PR 
where we introduced non-backward-compatible changes to zed and zq.

The data model description has been taken out of ZSON and moved into 
a dedicated specification and rationale added to doc/data-model/README.md.
The concept super-structured data is documented in the README and 
the Zed design motivated here.

doc/formats is now doc/data-model.  The primitive number types have been 
made more uniform and the error type turned into a complex type so 
errors can be wrapped.

ZJSON was changed quite a bit to better reflect the Zed model.  I think it's 
now finally right.

Since we are making breaking changes, the complex type codes in ZNG 
have been changed so the map type can be moved next to record, array, 
and set.

ZSON changes include: % sigil for enums, angle brackets for type values,
and the removal of bstring.

cmd/zed/README was deleted as it was never completed and will be 
replaced by ongoing work in zed/docs.