replace deprecated concept of "unset" with null
This commit replaces the deprecated concept of unset with Zed null.
Zed does not have unset record fields as a record with an unset
field is simply a different record type without that field.

We also removed the null check from zed.AppendInt and moved it
into zcode where it belongs.