Enable streaming stats on query endpoint
Stats were not getting sent on a periodic basis for the query endpoint.
This was causing long running queries on a service to time-out when
served behind an AWS load balancer: The load balancer would see no data
getting transmitted and time-out the connection.

Add a time Ticker in the query handler and also flush the response data
when a control message is sent.