add expr.Context and change Evaluator interface to take context (#3368)
This commit changes the Eval() interface to expr.Evaluator to
add an "expression context" representing the state of a locally
executing flowgraph.  This includes references to named values
via a forthcoming let/with capability of Zed flowgraphs along
with a means to allocate new Zed values inside of the context
so they may be easily freed to a pool when no longer needed
downstream.  Some scaffolding for let and ast.Scope has been put in
for a subsequent PR to finish these bits.

zbuf.Batch now has a method to carry its expression context.
This is important as flowgraph concurrency requires that batches and
context be carried together.  Since a batch is a bunch of values
and the context allows the creation of new values in the same
batch, this design provides a foundation for more efficient batch
allocations in the future.  (For example, batches could be wrapped
so new values in a new batch could refer to byte slices in the
underlying batch and the top batch would ref the underlying batch.)

Also as part of this change, we reworked the convention in
expr.Evaluator.Eval() to return all Zed runtime erros as Zed
values and only unrecoverable system errors (e.g., i/o errors
or corrupt data errors) as Go errors.  This can cause uncaught
panics to occur from proc goroutines but a soon-to-be-merged PR
will add recover() to each proc goroutine.

Missing values are now generally not ignored in group-bys, cut,
put etc.  If you want to ignore them in these context,
you can wrap the underlying expressions in the new quiet() function,
where quiet turns error("missing") into error("quiet") and quiet
errors *are ignored*.  That says, yield ignores top-levelmissing
values.

We also updated many places in the code to use *zed.Value pointers
instead of zed.Values.  The pick proc has been removed.

proc.Function has been replaced with expr.Applier and various cleanup
made to appliers.