eliminate driver and proc warnings channel (#3390)
This commit eliminates the driver package and drops the warning
channel from proc.Context.  Warnings are now delivered via
top-level error results in the dataflow output.  We will add a WarningsFilter
zio.Reader in a subsequent PR to filter Zed warnings and send them 
to stderr (zq) or to the warnings metadata (endpoint).

Queries are now run from a simpler API in package runtime, which
returns a query object that is a puller and is easily turned into
a zio.Reader or the new interface zbuf.ProgressReader.  Progress updates
are now more cleanly handled through this interface.

We also fixed the compiler to properly return meters for file system
scanners not just lakes and simplified the interface for retrieving
a meter for the flowgraph.

The logic to teardown a flowgraph is now all in proc.  In particular,
proc.Mux and the new type proc.Latcher (for single-output graphs)
both tear down the flowgraph at EOS and convert the double-EOS batch
protocol to a single-EOS semantic.  We also added proc.Catcher to
implement recover() to provide an easy means to catch panics that
can be raised in the expr package when encountering fatal errors
(like system I/O problems or corrupt ZNG data).  A subsequent PR 
will add a Closer to the Query return value so the caller call close it 
without reading to the end or canceling context.

Since the driver interface is gone, we needed a way to signal channel
completions.  This is accomplished by sending a proc.EndOfChannel
batch that is otherwise an empty batch but is detected by the
query endpoint in the server so it can send the corresponding events
to a client.

To carry control messages through the zio.Reader interface, we create
zbuf.Control, which implements the error interface, so control messages
including progress updates (zbuf.Progress) can be "read" for a zio.Reader.

As suggested by Noah, this commit removes the ctrl query parameter
from the /query endpoint so that control messages are always included.
To filter control when not wanted, zbuf.NoControl() wraps a zio.Reader
and filters all control messages.  The lake api Query endpoint automatically 
wraps a NoControl reader while the QueryWithControl endpoint is available 
to receive the control message stream.