fix scoping model and update const implementation (#3415)
This commit makes the long-wanted change to the compiler to handle consts
properly.  ConstProc and TypeProc are now gone and const statements can
appear at the front of any sequential flowgraph segment and are carried
in ast.Sequential.

The DAG no longer depends on the AST of zed values and types.  Instead,
they are encoded as ZSON strings.  The DAG also knows nothing about consts
as the semantic pass now performs constant propagation and all consts
are turned into DAG literals.

These changes were necessary at this time because without them, scoped
consts were getting hairy in compiler/kernel and the AsFilter() method
on zbuf.Filter needed to retain the table of consts.  By having the
semantic pass resolve all consts, this is all simpler and will work
well over the network for distributed queries.

To support dynamic type name resolution, we added the typename() function,
which takes a string type name and returns the zed type if it exists.
When named types are referred to and can't be resolved at compile time,
the semantic pass automatically inserts a call to typename().

This scoping work tees up the forthcoming PR on scoped variables
so subgraphs can refer to values from the outer "this" when traversing
new values for "this".