update threaded operators to handle batching protocol
This commit has a number of changes to accommodate the batching protocol
in various threaded operators including switch, exprswitch, combine,
and merge.  For merge and combine, there's no need to detect double-EOS
but rather convey double-EOS through the fan-in of nodes where we need
to block paths that have sent EOS until all of the paths have sent EOS,
then resume all paths at the same time.  Doing this in combination with
a deadlock-free done protocol was a bit tricky, but we have clarified
the Done contract to say if an operator initiates a Done it must immediately
call Pull until receiving EOS then send exactly one EOS downstream and
resume as if initialized.

We added an explicit merge operator to the Zed language.  Previously, merge
was used only by the compiler on lakes but having a merge operator in
Zed is by itself useful but also makes for easier testing.  This will be
undocumented for now.

These changes exposed a bug in the "from" operator where a race between
EOF of the scanner and a Done coming upstream could result in a file
being erroneously closed twice.

We also added various catchers near the top of goroutines.  We might want
to make this even a bit tighter by adding the defer/recover calls right
at the point of goroutine creation.  See issue #3420.