language improvements (drop semicolon, case keyword, match to search) (#3422)
This commit makes a number of improvements to the language for
clarity and consistency. Semicolons are no longer used anywhere.
Instead, new instroduced keywords to provide needed disambiguation:
"pool" (inside of a multi-way from) and "case" (for switch cases).

The new keywords provide familiarity and while they increase typing
a bit, they are used in contexts where you would be "editing code"
compared to "typing searches".

We also changed the "match()" expression syntax to "search()" as this
is clearer in intent.  This change was made in combination with the
fact that swithc case expressions are now always expressions whereas
before one style was boolean search and the other style was expression.
This just created confusion for me.  So, now you assume expression
syntax and where you want boolean search, you just wrap the text
with search().