lake.Writer: Use CopyFrom to copy values
lake.Writer was using value.Copy to copy a value bytes into the values
slice. This was resulting in an extra burden for allocations for every
added value, since a new byte slice had to be allocated. Instead use
value.CopyFrom which will attempt to use already allocated bytes slices.