update ZSON string escaping, delete bstring, simplify zeekio
As part of the final updates to the Zed formats, this commit
gets rid of the bstring type and updates the ZSON escaping
rules.  We now convert Zeek TSV strings, with their \x escapes,
into normal UTF-8 Zed strings.  Where an \x escape results in
illegal code points, we escape the the backslash before the x
just as the Zeek does for JSON logs.

Also, with this change it became clear that tzngio was no longer
needed and was consolidated into zeekio.  tableio and textio now
use zeekio.FormatValue().