add scopes to dataflow model (#3441)
This commit adds the concept of a scope to the dataflow model.  A scope
allows a subgraph in an inner scope to perform computation on values that
are derived from the sequence of values in the outer scope.  In this way,
the generated values run to completion in the scope and the scope yields
it results to the outer scope.  A scope can only exist in the context of
a traversal operator like "over" ("into" is coming soon).

A scope also allows the binding of variables (via a "let" operator)
to outer-scope values that appear in the inner scope as so named.
This is accomplished by modifying each batch that enters the scope
with the variables added by the scope.  Scopes can be nested and
each level of nesting pushes any new variables on top of any existing
scoped variables and popping them off the batch on exit from that scope.

To accommodate this new model, the double-EOS protocol is now restricted
to run only within a scope and batches that come out of a scope adhere
to a sinlge-EOS protocol.  We call each single-EOS-terminated sequence
of batches, terminated by a double-EOS, a platoon.

In making these changes, we realized we needed to simplify the semantics
of fanout operators (split/switch) and fan-in operators (combined/merge)
while also eliminating the Done method on proc.Interface by having a
done Boolean passed as a parameter to Pull.  proc.Interface is now
entirely subsumed by zbuf.Puller.  EOS and Done events are now always
resolved by a go-channel select on an inbound done and an outbound write
thus resolving races between sending EOS downstream and propagating done indications upstream.