add structured errors, change ZSON enum syntax
This commit adds structured errors to Zed as part of the Dec 2021
changes to the Zed formats.  Since a parser ambiguity arose that was
solved by the new enum syntax, we rolled the enum changes into this PR
as well.

Since bstring and stringy errors are gone, we got rid of the IsStringy
pattern and replaced it everywhere with a compare to type string.

The diffs are somewhat verbose because the change requires that errors
are now allocated from a type context so many more entities need a zctx.

Closes #3320