move all formatting out of package zed
This commit moves the last bits of formatting out of package zed.
This was an ancient vestige of the original code before we had
all the serialization formats figured out.  At this point, you
should a specific serialization format to get text formatting
of Zed data.

We also moved the ZNG validation logic out of zed and into zngio.

The zng.Type interface now consists of just ID() and Kind()
methods.  The Kind() method is new and return a string indicating
"primitive", "record", "set", etc.  This will be useful in
data shaping and introspection and we should add a kind()
function to Zed subsequent to this commit.

Finally, these changes exposed a number of bugs that have been fixed.