simplify ZST reassembly maps (#3488)
Now that ZST handles any type of top-level value, we simplified
the reassembly section that defines the reassembly maps.
This is a baby step toward a larger PR that will simplify the
column maps and clean up how the column maps are built.

We updated the ZST format doc to reflect the changes and the
new dig tooling and adding nulls to the tests as the rebase onto
zng-2022 fixed the bug with nulls in zst.  There is still a bug
with null container types but this will get fixed with an
anticipated refactor of the column maps.

We also fixed a bug with the handling of sets where set type instead
of the element type was passed to the column constructor.