unify trailers and add dev dig commands and simplify ZST
These two PRs  (#3486 and #3488) are being squashed into one commit 
after being approved separately.  They are committed together because they 
are interdependent.

The first PR (#3486) unifies trailers and adds dev dig commands.

This commit unifies the ZNG trailer structure used by ZST and
where the section layout is generically shared.  We also added
tooling as a new command "zed dev dig" to generically explore
trailered files.

"zed dev zngdump" is now called "zed dev dig frames" and now
outputs the frame metadata as Zed instead of printed text.

The "zst inspect" was removed since its functionality is now
subsumed by "zed dev dig".

Closes #3470

The second PR (#3488) simplifies ZST reassembly maps.

Now that ZST handles any type of top-level value, we simplified
the reassembly section that defines the reassembly maps.
This is a baby step toward a larger PR that will simplify the
column maps and clean up how the column maps are built.

We updated the ZST format doc to reflect the changes and the
new dig tooling and adding nulls to the tests as the rebase onto
zng-2022 fixed the bug with nulls in zst.  There is still a bug
with null container types but this will get fixed with an
anticipated refactor of the column maps.

We also fixed a bug with the handling of sets where set type instead
of the element type was passed to the column constructor.