add "with clause" to record literals (#3517)
This commit extends record expressions to be able to create new
records from a template record and additional fields, e.g., {r with x:1}
would give {a:1,b:2,x:1} if r equaled {a:1,b:2}. The template can be
any expression. This will be particularly useful in an inner scope
to join the outer this with inner-traversed values.