collect: ensure correct selector for type (#3531)
With canonical unions the collect aggregation was not deriving the
correct selector index for a union of values since the type order may be
rearranged when the zed.Context creates the union type. Fix this and add
a lookup table to zed.TypeUnion so it's easier to find a type's
selector.