fig bug unmarshaling zed.Value (#3535)
This commit fixes a bug where the unmarshaler would utilize the bytes
buffer of the zed.Value being decoded when returning an unmarshaled
zed.Value.  This can cause corruption to that zed.Value whenever the
original byte slice is changed as ownership is unclear in this scenario.
The fix is to copy the needed bytes from the incoming Zed value when
unmarshaling zed.Values.

This code path is not used much but manifested in the unmarshaling of
the commit journal when creating the scan snapshots.  It went
unnoticed because of the way the scan would get built: with all
of the buggy spans the same, every data object in the pool would be
merge-scanned for any query.  While  correct, this can obviously
cause really poor performance.