fix peeker bug and rework interactive behavior
This commit fixes a bug introduced in #3375 where ReadAtLeast()
could be called with a bigger request than the buffer provided.
The design intent behind Peeker is for parsers that need to look
ahead to have a convenient and efficient API for issuing large
reads but peeking ahead tiny amounts.  The design intent was not
to cause tiny reads to percolate through the syscall interface.

The "slow drip" use case is not clarified in #3375 but presumably
has to do with receiving stats updates by the Go client library
during a long-running server query or load operation.  In this case,
the updates accumulate slowly and are not returned by the Peeker
thus defeating the purposes of stats update.

Since the design change in #3375 interferes with the more common
use case of lake scans or zq issuing bulk reads, the fix here
is to condition interactive buffering on an option that can be
enabled by the client query library.  And while we're making this
change, the bug identified above has been fixed.