make default LZ buffer size much bigger
This commit changes the default LZ buffer size in the ZNG writer
from 16KB to 500KB leading to a speedup of 5X (in combination
with the checked-in buffering improvements) for Boyer-Moore string
search.  The compressed ZNG files are materially smaller.

I tried 1MB too but it didn't seem to make a difference over 500KB.