reorganize docs
This commit udate the docs directory with a simpler to understand
arrangement.  The reference docs in docs/zq/reference.md is mostly
complete while a subsequent PR will clean up the mess above.

Some markdown tests have been disabled with the pattern
mdtest-command-issue-xxx and issues put in backlog so they should
soon be fixed and merged.