rename "ndjson" format to "json" for -f and -i
The -f and -i format selection flags for Zed tools recognize both "json"
and "ndjson".  While the exact behavior for the "json" format is not
easy to describe, "-f json" is roughly equivalent to "-f ndjson"
together with "summarize collect(this) | yield this" while "-i json" is
roughly equivalent to "-i ndjson" together with "over this".  These Zed
language constructs weren't available when the "json" was originally
implemented, but now that they are, its specialzed behavior unneeded.

Simplify by renaming the "ndjson" format to "json".  More specifically,

1. Give "json" the current behavior of "ndjson" for the -f, -i, and
   zio/anyio.

2. Remove "ndjson" as a format recognized by -f, -i, and zio/anyio.

3. Remove zio/ndjsonio.

4. Replace zio/jsonio.WriterOpts and its ForceArray option with
   zio/zjsonio.ArrayWriter (for application/json HTTP responses).