tidy up the ksuid function
This commit updates the ksuid() function by allowing it to create
new KSUIDs (when called with no args) and to convert both ways from
bytes to string and from string to bytes.

Docs for ksuid will be updated in the docs-update branch and that PR
will be put up soon.

Closes #3564