change filter to search/where and add implied yield (#3620)
This commit changes the "filter" keyword to where and also allows
"search" to explicitly cause a parse of a keyword boolean search.
To do this, we changed ast.Filter to ast.OpExpr, which represents
an unresolved expression that requires semantic analysis to
determine if it is a boolean search, an implied yield, or an
simply function-call style aggregation.

Though "where/search" are now operator keywords, the filter operator
is still the implementation of these concepts so dag.Filter
remains and there is no dag.Where.

More thorough documentation for all of this will be soon merged
as part of issue #3604.  The stuff in where.md is draft-y and
will move around.