remove npm/build/index.js
The comment at the top of npm/build/index.js states that it mimics the
build recipe in Makefile but is written in JavaScript so it works on
Unix and Windows.  Bash and GNU Make are now easy to install on Windows
and available on the GitHub-hosted Windows runners for GitHub Actions,
making npm/build/index.js unnecessary.  Remove it and use "make build"
in package.json.