differentiate "in" and search with "grep" function
This commit finishes the work to separate containment checking
with "in" from nested string searches with a new function "grep".
It also cleans up the parse tree to fully disambiguate between
a "where operator" that uses a boolean expression and a
"search operator" that uses search syntax.  We do so by adding
AST nodes that are present only in searches.  We updated zfmt
with a check to determine whether an OpExpr is a "where" or a
"search" using these new AST nodes and formatting the canonical
form accordingly.

Since grep subsumes the "matches" operator for most cases, we
removed "matches" but we wanted to still allow regexp compare
in expression context so we revided the old ~ operator to provide
for a regexp comparison where the RHS must be a regexp.

This will all be more thoroughly documented in the forthcoming
zed-language PR.